# Conflicts:
#	KreemMachine/KreemMachineLibrary/Services/UserService.cs
#	KreemMachine/KreemMachineUI/MainWindow.xaml.cs

+++ KreemMachine/KreemMachineUI/MainWindow.xaml.cs
@@ -131,7 +131,8 @@ namespace KreemMachine
             RefreshProductsTableTimer.Elapsed += (sender, e) => Dispatcher.Invoke(() => RefreshProductsTable());
             RefreshDepartmentTableTimer.Elapsed += (sender, e) => Dispatcher.Invoke(() => RefreshDepartmentTable());
 
-            AllDepartmentsListBox.ItemsSource = departmentService.GetAll();
+            AllDepartmentsListBox.ItemsSource = departmentService.GetAllViewable();
+
         }
 
         private void SearchTextBox_KeyUp(object sender, KeyEventArgs e)
@@ -712,7 +713,7 @@ namespace KreemMachine
 
         private void StockStatisticsTab_Selected(object sender, RoutedEventArgs e)
         {
-            var departments = productServices.GetAllDepartments();
+            var departments = departmentService.GetAllViewable();
             cbxStockCategory.ItemsSource = departments;
             cbxStockCategory.SelectedItem = departments[0];
             cbxStockCategory.DisplayMemberPath = "Name";
@@ -732,7 +733,7 @@ namespace KreemMachine
 
         #endregion
 
-        #region Department 
+        #region Departments
 
         private void btnAddDepartment_Click(object sender, RoutedEventArgs e)
         {
@@ -743,10 +744,20 @@ namespace KreemMachine
 
         private void btnDeleteDepartment_Click(object sender, RoutedEventArgs e)
         {
-            var button = sender as Button;
-            var department = button.DataContext as Department;
+            RefreshDepartmentTableTimer.Stop();
+            if (MessageBox.Show("You are about to remove this department. Continue?", "Delete department", MessageBoxButton.YesNo, MessageBoxImage.Question) == MessageBoxResult.Yes)
+            {
+                var button = sender as Button;
+                var department = button.DataContext as Department;
 
-            departmentService.Delete(department);
+                int i = departmentService.Delete(department);
+                if (i == 1)
+                {
+                    SearchDepartmentsBar.Text = null;
+                    DepartmentTab_Selected(sender, e);
+                }
+                RefreshDepartmentTableTimer.Start();
+            }
         }
 
         private void btnEditDepartment_Click_1(object sender, RoutedEventArgs e)
@@ -771,9 +782,22 @@ namespace KreemMachine
 
         private void RefreshDepartmentTable() {
             AllDepartmentsListBox.ItemsSource = null;
-            AllDepartmentsListBox.ItemsSource = departmentService.GetAll();
+            AllDepartmentsListBox.ItemsSource = departmentService.GetAllViewable();
         }
 
+        private void SearchDepartmentsBar_KeyUp(object sender, KeyEventArgs e)
+        {
+            if (!string.IsNullOrWhiteSpace(SearchDepartmentsBar.Text))
+            {
+                RefreshDepartmentTableTimer.Stop();
+                AllDepartmentsListBox.ItemsSource = departmentService.FilterDepartments(SearchDepartmentsBar.Text);
+            }
+
+            else
+            {
+                DepartmentTab_Selected(sender, e);
+            }
+        }
         #endregion
 
     }
